requirements.txt
fastapi
uvicorn
pandas
snowflake-connector-python
python-dotenv